requirements.txt
altair==6.0.0
attrs==25.4.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.62.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jpype1==1.6.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
konlpy==0.6.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
narwhals==2.18.0
numpy==2.4.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pillow==12.1.1
plotly==6.6.0
protobuf==6.33.5
pyarrow==23.0.1
pycparser==3.0
pydeck==0.9.1
Pygments==2.19.2
PyMySQL==1.1.2
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.5
rich==13.9.4
rpds-py==0.30.0
selenium==4.18.1
setuptools==80.10.2
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
webdriver-manager==4.0.2
wheel==0.46.3
wordcloud==1.9.6
wsproto==1.3.2
xlrd==2.0.2
